refactor: APIコントローラをcontrollers/api配下に分離 (#213)
- JSON専用コントローラをApi::名前空間に移動
- APIエンドポイントを/api/...に統一
- JavaScript側の呼び出しURLを更新
- テストのルートヘルパーを更新
