+++ app/controllers/api/base_controller.rb
@@ -1,0 +1,6 @@
+# app/controllers/api/base_controller.rb
+module Api
+  class BaseController < ApplicationController
+    before_action :authenticate_user!
+  end
+end
