+++ app/controllers/api/goal_points_controller.rb
@@ -1,0 +1,34 @@
+# app/controllers/api/goal_points_controller.rb
+module Api
+  class GoalPointsController < BaseController
+    include Recalculable
+
+    before_action :set_plan
+
+    def update
+      goal_point = @plan.goal_point || @plan.build_goal_point
+      goal_point.update!(goal_point_params)
+
+      # 帰宅地点変更後に route → schedule を再計算
+      recalculate_route_and_schedule!(@plan)
+
+      render json: {
+        address: goal_point.address,
+        lat: goal_point.lat,
+        lng: goal_point.lng
+      }, status: :ok
+    rescue ActiveRecord::RecordInvalid => e
+      render json: { message: "帰宅地点の更新に失敗しました", details: e.record.errors.full_messages }, status: :unprocessable_entity
+    end
+
+    private
+
+    def set_plan
+      @plan = current_user.plans.find(params[:plan_id])
+    end
+
+    def goal_point_params
+      params.require(:goal_point).permit(:address, :lat, :lng)
+    end
+  end
+end
